facture et rapport good

+++ src/main/java/com/diagnostic_immobilier_backend/demo/repository/FactureRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FactureRepository extends JpaRepository <Facture,Integer> {
+public interface FactureRepository extends JpaRepository <Facture,Long> {
 
 }
